Create Hadoop in Big Data Analytcs
There is a great deal of complexity involved in big data, including structured, semi-structured, and unstructured data, which makes it very difficult to analyse and process. The data had to be processed at a rapid speed, so Hadoop was used because of its storage capabilities and processing capabilities. Additionally, Hadoop is open-source software. As for the cost, it is a good choice for business solutions. Since it implements simple programming models, this framework has been popular in recent years because it permits distributed processing of enormous data sets using crosswise clusters of computers.